+++ src/main/java/dev/marksman/collectionviews/CopyOptimizeCheck.java
@@ -1,0 +1,5 @@
+package dev.marksman.collectionviews;
+
+interface CopyOptimizeCheck {
+    boolean shouldNotMakeCopy();
+}