+++ src/main/java/dev/marksman/collectionviews/ImmutableVectors.java
@@ -5,6 +5,7 @@ import com.jnape.palatable.lambda.adt.hlist.Tuple2;
 import com.jnape.palatable.lambda.functions.Fn1;
 import com.jnape.palatable.lambda.functions.Fn2;
 import com.jnape.palatable.lambda.functions.builtin.fn2.Take;
+import dev.marksman.enhancediterables.ImmutableFiniteIterable;
 import dev.marksman.enhancediterables.ImmutableNonEmptyFiniteIterable;
 import dev.marksman.enhancediterables.NonEmptyIterable;
 
@@ -24,9 +25,15 @@ import static dev.marksman.collectionviews.ImmutableCrossJoinVector.immutableCro
 import static dev.marksman.collectionviews.ImmutableReverseVector.immutableReverseVector;
 import static dev.marksman.collectionviews.ImmutableVectorZip.immutableVectorZip;
 import static dev.marksman.collectionviews.MapperChain.mapperChain;
-import static dev.marksman.collectionviews.Validation.*;
+import static dev.marksman.collectionviews.Validation.validateCopyFrom;
+import static dev.marksman.collectionviews.Validation.validateDrop;
+import static dev.marksman.collectionviews.Validation.validateFill;
+import static dev.marksman.collectionviews.Validation.validateNonEmptyFill;
+import static dev.marksman.collectionviews.Validation.validateSlice;
+import static dev.marksman.collectionviews.Validation.validateTake;
 import static dev.marksman.collectionviews.Vector.empty;
 import static dev.marksman.collectionviews.VectorSlicing.sliceImpl;
+import static dev.marksman.enhancediterables.FiniteIterable.emptyFiniteIterable;
 
 final class ImmutableVectors {
 
@@ -48,7 +55,7 @@ final class ImmutableVectors {
 
     static <A> ImmutableVector<A> copyFrom(Iterable<A> source) {
         Objects.requireNonNull(source);
-        if (source instanceof ImmutableVector<?> && Util.isPrimitive(source)) {
+        if (source instanceof ImmutableVector<?> && Util.shouldNotMakeCopy(source)) {
             return (ImmutableVector<A>) source;
         } else if (!source.iterator().hasNext()) {
             return Vectors.empty();
@@ -63,7 +70,7 @@ final class ImmutableVectors {
         if (maxCount == 0) {
             return Vectors.empty();
         }
-        if (source instanceof ImmutableVector<?> && Util.isPrimitive(source)) {
+        if (source instanceof ImmutableVector<?> && Util.shouldNotMakeCopy(source)) {
             return ((ImmutableVector<A>) source).take(maxCount);
         } else {
             return copyFrom(Take.take(maxCount, source));
@@ -72,7 +79,7 @@ final class ImmutableVectors {
 
     static <A> ImmutableVector<A> copySliceFrom(int startIndex, int endIndexExclusive, Iterable<A> source) {
         validateSlice(startIndex, endIndexExclusive, source);
-        if (source instanceof ImmutableVector<?> && Util.isPrimitive(source)) {
+        if (source instanceof ImmutableVector<?> && Util.shouldNotMakeCopy(source)) {
             return ((ImmutableVector<A>) source).slice(startIndex, endIndexExclusive);
         } else {
             return Vectors.sliceFromIterable(startIndex, endIndexExclusive, source).toImmutable();
@@ -336,6 +343,20 @@ final class ImmutableVectors {
         } else {
             return sliceImpl(ImmutableVectorSlice::immutableVectorSlice, source.size(), () -> source,
                     startIndex, requestedSize);
+        }
+    }
+
+    static <A> ImmutableFiniteIterable<? extends ImmutableNonEmptyVector<A>> slide(int k, ImmutableVector<A> source) {
+        if (k < 1) {
+            throw new IllegalArgumentException("k must be >= 1");
+        }
+        int size = source.size();
+        if (size == 0) {
+            return emptyFiniteIterable();
+        } else {
+            final int windowSize = Math.min(k, size);
+            final int maxIndex = 1 + size - windowSize;
+            return () -> new ImmutableVectorSlidingIterator<>(source, windowSize, maxIndex);
         }
     }