+++ CHANGELOG.md
@@ -6,6 +6,30 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## [Unreleased]
 No changes yet
 
+## [1.2.3] - 2020-07-19
+### Changes
+- `slide` now returns slices of the source `Vector`
+
+## [1.2.2] - 2020-04-18
+### Changes
+- Upgrade [enhanced-iterables](https://github.com/kschuetz/enhanced-iterables) to version 1.2.0
+
+## [1.2.1] - 2020-04-08
+### Changes
+- `ImmutableVector#cross` optimizations
+
+## [1.2.0] - 2020-04-03
+### Added
+- `SetBuilder`
+
+### Changes
+- `distinct` method on `Set`s now return the `Set` itself 
+
+## [1.1.2] - 2020-02-18
+### Changes
+- Upgrade [lambda](https://github.com/palatable/lambda) to version 5.2.0
+- Upgrade [enhanced-iterables](https://github.com/kschuetz/enhanced-iterables) to version 1.1.1
+
 ## [1.1.1] - 2020-01-08
 ### Added
 - `nonEmptyCopyFrom`
@@ -38,7 +62,12 @@ No changes yet
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/kschuetz/collection-views/compare/collection-views-1.1.1...HEAD
+[Unreleased]: https://github.com/kschuetz/collection-views/compare/collection-views-1.2.3...HEAD
+[1.2.3]: https://github.com/kschuetz/collection-views/compare/collection-views-1.2.2...collection-views-1.2.3
+[1.2.2]: https://github.com/kschuetz/collection-views/compare/collection-views-1.2.1...collection-views-1.2.2
+[1.2.1]: https://github.com/kschuetz/collection-views/compare/collection-views-1.2.0...collection-views-1.2.1
+[1.2.0]: https://github.com/kschuetz/collection-views/compare/collection-views-1.1.2...collection-views-1.2.0
+[1.1.2]: https://github.com/kschuetz/collection-views/compare/collection-views-1.1.1...collection-views-1.1.2
 [1.1.1]: https://github.com/kschuetz/collection-views/compare/collection-views-1.1.0...collection-views-1.1.1
 [1.1.0]: https://github.com/kschuetz/collection-views/compare/collection-views-1.0.4...collection-views-1.1.0
 [1.0.4]: https://github.com/kschuetz/collection-views/compare/collection-views-1.0.3...collection-views-1.0.4